# app/api/menu-engineering/openapi.yaml
openapi: 3.0.3
info:
  title: Carmen ERP - Menu Engineering API
  description: |
    Comprehensive Menu Engineering API for sales data integration, performance analysis, 
    and recipe optimization using Boston Matrix methodology.
    
    ## Features
    - **POS Integration**: Import and synchronize sales data from Point of Sale systems
    - **Menu Analysis**: Comprehensive performance analysis using Boston Matrix (Stars, Plowhorses, Puzzles, Dogs)
    - **AI Recommendations**: Intelligent optimization suggestions based on performance data
    - **Real-time Costing**: Dynamic recipe cost calculations with current ingredient prices
    - **Cost Monitoring**: Active alerts and threshold monitoring for cost variations
    - **Performance Metrics**: Detailed analytics for recipe performance and trends
    
    ## Security
    All endpoints require authentication via JWT or Keycloak tokens and implement:
    - Role-based access control (RBAC)
    - Rate limiting and DDoS protection
    - Input validation and sanitization
    - Comprehensive audit logging
    - Security headers and CORS policies
    
    ## Rate Limits
    - Sales Import: 50 requests/hour
    - POS Sync: 20 requests/hour
    - Analytics: 100-300 requests/hour
    - Recipe APIs: 200-500 requests/hour
    
  version: 1.0.0
  termsOfService: https://carmen-erp.com/terms
  contact:
    name: Carmen ERP API Support
    url: https://carmen-erp.com/support
    email: api-support@carmen-erp.com
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - KeycloakAuth: []
paths:
...
  /menu-engineering/recommendations/{recipeId}:
    get:
      tags:
        - Menu Analysis
      summary: Get AI recommendations for specific recipe
      description: |
        Get AI-powered optimization recommendations for a specific recipe based on performance analysis.
      operationId: getRecipeRecommendations
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Recipe identifier
        - name: priority
          in: query
          schema:
            type: string
            enum: [high, medium, low, all]
            default: all
          description: Filter by recommendation priority
        - name: impact
          in: query
          schema:
            type: string
            enum: [high, medium, low, all]
            default: all
          description: Filter by expected impact
        - name: effort
          in: query
          schema:
            type: string
            enum: [high, medium, low, all]
            default: all
          description: Filter by implementation effort
        - name: maxRecommendations
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
          description: Maximum number of recommendations
        - name: includeActionSteps
          in: query
          schema:
            type: boolean
            default: true
          description: Include detailed action steps
        - name: includeImpactEstimates
          in: query
          schema:
            type: boolean
            default: false
          description: Include estimated impact on revenue/profit
      responses:
        '200':
          description: Recommendations retrieved successfully
          headers:
            X-Recipe-Id:
              schema:
                type: string
              description: Recipe identifier
            X-Recommendation-Count:
              schema:
                type: string
              description: Number of recommendations returned
            X-High-Priority-Count:
              schema:
                type: string
              description: Number of high-priority recommendations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RecommendationsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ====== RECIPE ENHANCEMENT ENDPOINTS ======
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    KeycloakAuth:
      type: oauth2
      description: Keycloak OAuth2 authentication
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.carmen-erp.com/auth/realms/carmen/protocol/openid-connect/auth
          tokenUrl: https://keycloak.carmen-erp.com/auth/realms/carmen/protocol/openid-connect/token
          scopes:
            read:menu_engineering: Read menu engineering data
            write:menu_engineering: Write menu engineering data
            read:recipes: Read recipe data
            write:recipes: Write recipe data
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          description: Response data (present on success)
        error:
          type: string
          description: Error message (present on failure)
        message:
          type: string
          description: Additional message or context
        details:
          description: Additional error details or validation errors
        metadata:
          $ref: '#/components/schemas/ApiMetadata'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    ApiMetadata:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        version:
          type: string
        cacheTtl:
          type: integer
          description: Cache TTL in seconds
        generatedAt:
          type: string
          format: date-time
    PaginationInfo:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: decimal
          description: Monetary amount
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code
...
    MenuRecommendation:
      type: object
      required:
        - recipeId
        - recipeName
        - currentClassification
        - type
        - priority
        - impact
        - effort
        - title
        - description
      properties:
        id:
          type: string
        recipeId:
          type: string
          format: uuid
        recipeName:
          type: string
        currentClassification:
          type: string
          enum: [STAR, PLOWHORSES, PUZZLE, DOG]
        type:
          type: string
          enum: [promote, reposition, reprice, reformulate, remove, feature, bundle, investigate]
        priority:
          type: string
          enum: [high, medium, low]
        impact:
          type: string
          enum: [high, medium, low]
        effort:
          type: string
          enum: [high, medium, low]
        title:
          type: string
        description:
          type: string
        rationale:
          type: string
        expectedOutcome:
          type: string
        timeframe:
          type: string
        actionSteps:
          type: array
          items:
            type: string
        metrics:
          type: array
          items:
            type: string
        impactEstimates:
          type: object
          properties:
            revenueImpact:
              $ref: '#/components/schemas/Money'
            profitImpact:
              $ref: '#/components/schemas/Money'
        tags:
          type: array
          items:
            type: string
...
    RecommendationsResponse:
      type: object
      properties:
        recipeId:
          type: string
          format: uuid
        recipeName:
          type: string
        analysisDate:
          type: string
          format: date-time
        totalRecommendations:
          type: integer
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/MenuRecommendation'
        summary:
          type: object
          properties:
            highPriority:
              type: integer
            mediumPriority:
              type: integer
            lowPriority:
              type: integer
            quickWins:
              type: integer
            strategicInitiatives:
              type: integer

    # ====== RECIPE ENHANCEMENT SCHEMAS ======
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input parameters
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Authentication required"
...
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Resource not found"
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Internal server error"
...